Roave/BackwardCompatibilityCheck#767: Refactor, remove new EnumBased interface, treat Enums as just classes
