+++ src/CompareClasses.php
@@ -8,7 +8,6 @@ use Psl\Dict;
 use Psl\Regex;
 use Psl\Str;
 use Roave\BackwardCompatibility\DetectChanges\BCBreak\ClassBased\ClassBased;
-use Roave\BackwardCompatibility\DetectChanges\BCBreak\EnumBased\EnumBased;
 use Roave\BackwardCompatibility\DetectChanges\BCBreak\InterfaceBased\InterfaceBased;
 use Roave\BackwardCompatibility\DetectChanges\BCBreak\TraitBased\TraitBased;
 use Roave\BetterReflection\Reflection\ReflectionClass;
@@ -21,7 +20,6 @@ final class CompareClasses implements CompareApi
         private ClassBased $classBasedComparisons,
         private InterfaceBased $interfaceBasedComparisons,
         private TraitBased $traitBasedComparisons,
-        private EnumBased|null $enumBasedComparisons = null, // optional argument for now to avoid BC break
     ) {
     }
 
@@ -87,12 +85,6 @@ final class CompareClasses implements CompareApi
 
         if ($oldSymbol->isTrait()) {
             yield from ($this->traitBasedComparisons)($oldSymbol, $newClass);
-
-            return;
-        }
-
-        if ($oldSymbol->isEnum() && $this->enumBasedComparisons) {
-            yield from ($this->enumBasedComparisons)($oldSymbol, $newClass);
 
             return;
         }
