+++ test/unit/CompareClassesTest.php
@@ -10,7 +10,6 @@ use Roave\BackwardCompatibility\Change;
 use Roave\BackwardCompatibility\Changes;
 use Roave\BackwardCompatibility\CompareClasses;
 use Roave\BackwardCompatibility\DetectChanges\BCBreak\ClassBased\ClassBased;
-use Roave\BackwardCompatibility\DetectChanges\BCBreak\EnumBased\EnumBased;
 use Roave\BackwardCompatibility\DetectChanges\BCBreak\InterfaceBased\InterfaceBased;
 use Roave\BackwardCompatibility\DetectChanges\BCBreak\TraitBased\TraitBased;
 
@@ -28,9 +27,6 @@ final class CompareClassesTest extends TestCase
     /** @var TraitBased&MockObject */
     private TraitBased $traitBasedComparison;
 
-    /** @var EnumBased&MockObject */
-    private EnumBased $enumBasedComparison;
-
     private CompareClasses $compareClasses;
 
     public static function setUpBeforeClass(): void
@@ -45,12 +41,10 @@ final class CompareClassesTest extends TestCase
         $this->classBasedComparison     = $this->createMock(ClassBased::class);
         $this->interfaceBasedComparison = $this->createMock(InterfaceBased::class);
         $this->traitBasedComparison     = $this->createMock(TraitBased::class);
-        $this->enumBasedComparison      = $this->createMock(EnumBased::class);
         $this->compareClasses           = new CompareClasses(
             $this->classBasedComparison,
             $this->interfaceBasedComparison,
             $this->traitBasedComparison,
-            $this->enumBasedComparison,
         );
     }
 
@@ -59,7 +53,6 @@ final class CompareClassesTest extends TestCase
         $this->classBasedComparatorWillBeCalled();
         $this->interfaceBasedComparatorWillNotBeCalled();
         $this->traitBasedComparatorWillNotBeCalled();
-        $this->enumBasedComparatorWillNotBeCalled();
 
         Assertion::assertChangesEqual(
             Changes::fromList(Change::changed('class change', true)),
@@ -96,8 +89,6 @@ PHP,
         $this->classBasedComparatorWillBeCalled();
         $this->interfaceBasedComparatorWillNotBeCalled();
         $this->traitBasedComparatorWillNotBeCalled();
-        $this->enumBasedComparatorWillNotBeCalled();
-
 
         Assertion::assertChangesEqual(
             Changes::fromList(Change::changed('class change', true)),
@@ -130,8 +121,6 @@ PHP,
         $this->classBasedComparatorWillNotBeCalled();
         $this->interfaceBasedComparatorWillBeCalled();
         $this->traitBasedComparatorWillNotBeCalled();
-        $this->enumBasedComparatorWillNotBeCalled();
-
 
         Assertion::assertChangesEqual(
             Changes::fromList(Change::changed('interface change', true)),
@@ -147,9 +136,7 @@ PHP,
     {
         $this->classBasedComparatorWillNotBeCalled();
         $this->interfaceBasedComparatorWillNotBeCalled();
-        $this->enumBasedComparatorWillNotBeCalled();
         $this->traitBasedComparatorWillBeCalled();
-
 
         Assertion::assertChangesEqual(
             Changes::fromList(Change::changed('trait change', true)),
@@ -161,31 +148,11 @@ PHP,
         );
     }
 
-    public function testWillRunEnumComparators(): void
-    {
-        $this->classBasedComparatorWillNotBeCalled();
-        $this->interfaceBasedComparatorWillNotBeCalled();
-        $this->enumBasedComparatorWillBeCalled();
-        $this->traitBasedComparatorWillNotBeCalled();
-
-
-        Assertion::assertChangesEqual(
-            Changes::fromList(Change::changed('enum change', true)),
-            ($this->compareClasses)(
-                (self::$stringReflectorFactory)('<?php enum A {}'),
-                (self::$stringReflectorFactory)('<?php enum A {}'),
-                (self::$stringReflectorFactory)('<?php enum A {}'),
-            ),
-        );
-    }
-
     public function testAnonymousClassesAreFilteredOut(): void
     {
         $this->classBasedComparatorWillNotBeCalled();
         $this->interfaceBasedComparatorWillNotBeCalled();
         $this->traitBasedComparatorWillNotBeCalled();
-        $this->enumBasedComparatorWillNotBeCalled();
-
         Assertion::assertChangesEqual(
             Changes::empty(),
             ($this->compareClasses)(
@@ -240,8 +207,6 @@ PHP),
         $this->classBasedComparatorWillNotBeCalled();
         $this->interfaceBasedComparatorWillNotBeCalled();
         $this->traitBasedComparatorWillNotBeCalled();
-        $this->enumBasedComparatorWillNotBeCalled();
-
 
         Assertion::assertChangesEqual(
             Changes::fromList(Change::removed('Class A has been deleted', true)),
@@ -296,27 +261,10 @@ PHP),
             ->willReturn(Changes::fromList(Change::changed('trait change', true)));
     }
 
-    private function enumBasedComparatorWillBeCalled(): void
-    {
-        $this
-            ->enumBasedComparison
-            ->expects(self::atLeastOnce())
-            ->method('__invoke')
-            ->willReturn(Changes::fromList(Change::changed('enum change', true)));
-    }
-
     private function traitBasedComparatorWillNotBeCalled(): void
     {
         $this
             ->traitBasedComparison
-            ->expects(self::never())
-            ->method('__invoke');
-    }
-
-    private function enumBasedComparatorWillNotBeCalled(): void
-    {
-        $this
-            ->enumBasedComparison
             ->expects(self::never())
             ->method('__invoke');
     }
