+++ test/unit/DetectChanges/BCBreak/ClassBased/DummyEnum.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace RoaveTest\BackwardCompatibility\DetectChanges\BCBreak\ClassBased;
+
+enum DummyEnum
+{
+    case someCase;
+}
