--- test/unit/DetectChanges/BCBreak/EnumBased/DummyEnum.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace RoaveTest\BackwardCompatibility\DetectChanges\BCBreak\EnumBased;
-
-enum DummyEnum
-{
-    case someCase;
-}
